[MIN-133] Add a macro for orml traits implementation
